made blueprints table a reusable Table component, implemented on index page
